changed sticky and unsticky context menus to togglesticky and made tag resource as context menu cmd
